requirements.txt
PyQt5>=5.15.0
python-docx>=0.8.11
Pillow>=8.0.0
keyboard>=0.13.5 